Fix #696797, gracefully handle empty pidfile in logrotate
Under certain circumstances we could end up with an empty pidfile[1].
The logrotate script should handle that gracefully. We now also check
that the pidfile size is not zero before signaling nginx master.

[1] `nginx -t` creates/opens an empty pidfile without deleting it to
avoid posible race conditions. So if nginx is not running at the moment,
you will end up with an empty pidfile.

Relevant discusion:
http://mailman.nginx.org/pipermail/nginx/2008-May/005229.html