Fix #734139, log re-opening issue
The change of /var/log/nginx permissions to root:adm 0750 introduced
a serious log rotate issue.

According to the nginx documentation:
http://nginx.org/en/docs/control.html (Rotating Log-Files)

In order to rotate log files, they need to be renamed first. After that
USR1 signal should be sent to the master process. The master process
will then re-open all currently open log files and assign them an
unprivileged user under which the worker processes are running, as an
owner. After successful re-opening, the master process closes all open
files and *sends the message to worker process to ask them to re-open
files*.

The problem is that worker processes running as www-data cannot re-open
the log files as www-data does not have execute permission to
/var/log/nginx. The workers keep writing to /var/log/nginx/access.log.1
and after a second run of logrotate you will end up with corrupted log
files because of gzipping.

We now change the permissions to www-data:adm 0750.