Drop nginx-naxsi packages
Removing everything naxsi related from debian/

We have decided to remove the nginx-naxsi packages before jessie is
freezed.

Packaging naxsi is not trivial and, unfortunately, none of the
maintainers uses it. That's the reason nginx-naxsi is not in a good
shape and we are not feeling comfortable to release and support it.