Make sure signature stays the same in all nginx builds
NGX_HTTP_HEADERS is part of nginx signature. When a dyn
modules is loaded the signature of the module is compared
to the one of the nginx binary.

dyn modules are build from nginx-full, so in order to make
them loadable in other flavors we need to make sure all the
binaries share the same signature.

Gbp-Pq: Name 0002-Make-sure-signature-stays-the-same-in-all-nginx-buil.patch